Relocated the App.js to the components folder
